(tui) work on lua bindings and test cases
Fixes some flaws in the attribute- setting API so that tui_attr can be created
via a template. Also covers an issue with the context passed as first argument
to the callbacks not refering to the right window.

+++ src/shmif/tui/tui_lua.h
@@ -40,7 +40,7 @@
  */
 struct tui_lmeta {
 	struct tui_context* tui;
-	int href;
+	intptr_t href;
 
 /* pending subsegment requests and their respective lua references */
 	uint8_t pending_mask;
